CVsim Parameters

The main object of CVsim.jl is the CVsim_Parameters structure that is used for constructing the CV states and in state preparation and stabilization protocols. It is constructed using the Parameters.jl package and stores all the constants defining a particular simulation.

CVsim.CVsim_Parameters — Type

CVsim_Parameters( ; qvec_min = -8*√π, qvec_max = 8*√π, n_points = 300, 
                    b_prefered = "position", spin_number = 1//2, qudit_levels = 3 )

Structure storing essential parameters that define a given simulation

Inputs (optional):

qvec_min::Float64 : Minimum position in phase space.

qvec_max::Float64 : Maximum position in phase space.

n_points::Int : Number of points equidistantly distributed between qvec_min and qvec_max.

b_prefered::String : Preferred basis for state representation ("position" or "momentum").

spin_number::Rational{Int} : Spin number for a spin auxillary system.

qudit_levels::Int : Number for qudit levels for a qudit auxillary system.

Output:

A CVsim_Parameters structure containing these parameters as well as the essential bases (position, momentum, spin, qudit).