GKP state preparation scripts

CVsim.prepare_GKP — Function

prepare_GKP( P; method = "default", kwargs...)

Preparation of Gottesman-Kitaev-Preskill (GKP) states using various methods.

Inputs:

P::CVsim_Parameters : Parameter structure as defined in parameters.jl

method::String : Method to prepare a GKP state. Currently only "default", "Hastrup", "Albert", "Mensen" methods are implemented. See their definitions bellow.

kwargs : Other parameters specific to each method.

Output:

A GKP state as a Ket state or as an Operator depending on the method (density matrix)

CVsim.default_GKP_preparation — Function

default_GKP_preparation( P; μ = 0.0, Δ = 0.37, κ = 0.37, period = 2√π, parity = 0)

Preparation of the approximate GKP state using a sum of gaussians with an outer envelope

Inputs:

P::CVsim_Parameters : Parameter structure as defined in parameters.jl

μ::Float64 : Center of the big gaussian envelope

Δ::Float64 : Δ² - variance of Gaussian peaks in the GKP state

κ::Float64 : 1/κ² - variance of the Gaussian envelope

period::Float64 : Period of the θ function,

set 2√π for |0⟩L and |1⟩L states

set √π for |+⟩L and |-⟩L states

parity::Bool : Boolean that sets the parity of the state

set 0 for |0⟩L and |+⟩L

set 1 for |1⟩L and |-⟩L

n_peaks::Int : Half of the total number of Gaussian peaks that that compose the state

Output:

|GKP⟩ Ket vector

CVsim.hastrup_preparation — Function

hastrup_preparation( P; n_cycles = 2, κ_squeez = .37, σ_ground = 1.0)

Preparation of the approximate GKP state using an additional two-level quantum system and measurement-free preparation protocol as proposed by Hastrup et al. in DOI: 10.1038/s41534-020-00353-3

Warning: The returned state has a spin component. Inputs:

P::CVsim_Parameters : Parameter structure as defined in parameters.jl

n_cycles::Int : Number of cycles of preparation as defined in the paper. Suported numbers are 2, 3, 4.

κ_squeez::Float64 : Squeezing of the initial squeezed vacuum state

σ_ground::Float64 : Standard deviation of the vacuum state (gaussian state)

Output:

|0⟩L GKP Ket ⊗ spin Ket

CVsim.albert_preparation — Function

albert_preparation( P; Δ = 0.37, n_peaks = 51, GKP_1 = false)

Preparation of the approximate GKP state by summing a certain number of displaced squeezed states as reported by Albert et al. in Eq.(7.7a) from DOI: 10.1103/PhysRevA.97.032346

Inputs:

P::CVsim_Parameters : Parameter structure as defined in parameters.jl

Δ::Float64 : Δ² - variance of Gaussian peaks in the GKP state, and 1/Δ² - variance of the Gaussian envelope

n_peaks::Int : Number of peaks to initialize

GKP_1::Bool : if set to true |1⟩L will be initialized otherwise |0⟩L

Output:

|GKP⟩ Ket vector

CVsim.mensen_preparation — Function

mensen_preparation( P; μ = 0.0, Δ = 0.37, κ = 0.37, period = 2√π, logic = [[0],[0]])

Preparation of the approximate GKP state using the Jacobi θ function as as reported by Mensen et al. in Eq.(3.34) from DOI: 10.1103/PhysRevA.104.022408

To evaluate the Jacobi θ function we use the package Theta.jl from DOI: 10.2140/jsag.2021.11.41

Inputs:

P::CVsim_Parameters : Parameter structure as defined in parameters.jl

μ::Float64 : Center of the big gaussian envelope

Δ::Float64 : Δ² - variance of Gaussian peaks in the GKP state

κ::Float64 : 1/κ² - variance of the Gaussian envelope

period::Float64 : Period of the θ function,

set 2√π for |0⟩L and |1⟩L states

set √π for |+⟩L and |-⟩L states

logic::Vector{Vector{Int}} : Vector of two booleans which sets the order of the peaks

set [[0],[0]] for |0⟩L and |+⟩L

set [[0],[1]] for |1⟩L

set [[1],[0]] for |-⟩L

Output:

|GKP⟩ Ket vector