GKP state stabilization scripts

CVsim.stabilize_GKP — Function

stabilize_GKP( rho, P; method = "DeNeeve", kwargs...)

Stabilization of Gottesman-Kitaev-Preskill (GKP) states using various methods

Inputs:

rho::Operator{B,B} Operator (density matrix) defined in the basis B

P::CVsim_Parameters Parameter structure as defined in parameters.jl

method::String Method to prepare a GKP state. Default: "DeNeeve"

kwargs Other parameters specific to each method.

Output:

A GKP state as an Operator (density matrix)

stabilize_GKP( psi, P; method = "DeNeeve", kwargs...)

Stabilization of Gottesman-Kitaev-Preskill (GKP) states using various methods (for now only De Neeve's method is available).

Inputs:

psi::StateVector{B} State vector (Ket or Bra) defined in the basis B

P::CVsim_Parameters Parameter structure as defined in parameters.jl

method::String Method to prepare a GKP state. Default: "DeNeeve"

kwargs Other parameters specific to each method.

Output:

A GKP state as an Operator (density matrix)

CVsim.DeNeeve_stabilization — Function

DeNeeve_stabilization( rho, P; n_times = 1, ε = 2√π*0.045, α = √π, μ = 2√π*0.065, option = "", indices = 0, kraus = true)

Stabilization of Gottesman-Kitaev-Preskill (GKP) using the spin degree of freedom of an ion as reported by De Neeve et al. in DOI: 10.1038/s41567-021-01487-7

Inputs:

rho::Operator{B,B} Operator (density matrix) defined in the position/momentum basis B or a composite basis B with at least one position/momentum basis in it.

P::CVsim_Parameters Parameter structure as defined in parameters.jl

n_times::Int Number of stabilization rounds

ε, α, μ ::Float Three parametres as defined in the paper

option::String String specifying which can contain "q" or "p" for correcting only one of the quadratures. If none, both will be stabilized.

indices Index or vector of indices of subspaces that one would like to stabilize. If index = 0, then all the relevant subspaces will be subjected to n_cycles stabilization.

kraus::Bool Use Kraus operators instead of the unitaries and the spin degree of freedom. Default: true.

Output:

A GKP state as an Operator (density matrix)

CVsim.CampagneIbarcq_stabilization — Function

CampagneIbarcq_stabilization( rho, P; n_times = 1, l = 2√π, Δ = 0.37, option="all", indices = 0)

Stabilization of Gottesman-Kitaev-Preskill (GKP) using an auxillary transmon and consisting of a sharpening and trimming steps for each quadrature as reported by Campagne-Ibarcq et al. in DOI: 10.1038/s41586-020-2603-3

Inputs:

rho::Operator{B,B} Operator (density matrix) defined in the position/momentum basis B or a composite basis B with at least one position/momentum basis in it.

P::CVsim_Parameters Parameter structure as defined in parameters.jl

n_times::Int Number of stabilization rounds

l::Float Lattice constant of a square lattice encoding

Δ::Float Δ² - variance of Gaussian peaks in the GKP state 1/Δ² - variance of the Gaussian envelope

option::String String specifying which steps to execute: "qS" - only sharpening of q-quadrature "qST" - sharpening-trimming of q only Default: "all" which is equivallent to "qpST"

indices Index or vector of indices of subspaces that one would like to stabilize. If index = 0, then all the relevant subspaces will be subjected to n_cycles stabilization.

Output:

A GKP state as an Operator (density matrix)

CVsim.Royer_stabilization — Function

Royer_stabilization( rho, P; n_times = 1, l = 2√π, Δ = 0.37, option="sBs", indices = 0)

Stabilization of Gottesman-Kitaev-Preskill (GKP) using an auxillary qubit and either of the two methods Big-small-Big or small-Big-small as reported by Royer et al. in Eqs. (6b) and (6c), respectively, from DOI: 10.1038/s41586-020-2603-3

Inputs:

rho::Operator{B,B} Operator (density matrix) defined in the position/momentum basis B or a composite basis B with at least one position/momentum basis in it.

P::CVsim_Parameters Parameter structure as defined in parameters.jl

n_times::Int Number of stabilization rounds

l::Float Lattice constant of a square lattice encoding

Δ::Float Δ² - variance of Gaussian peaks in the GKP state 1/Δ² - variance of the Gaussian envelope

option::String String specifying which method to use: "BsB" - Big-small-Big (work well for Δ ≤ 0.25) "sBs" - small-Big-small (work better in general) The option can also contain "q" or "p" for correcting only one of the quadratures. If none, both will be stabilized. Default: "sBs"

indices Index or vector of indices of subspaces that one would like to stabilize. If index = 0, then all the relevant subspaces will be subjected to n_cycles stabilization.

Output:

A GKP state as an Operator (density matrix)

CVsim.Sivak_stabilization — Function

Sivak_stabilization( rho, P; n_times = 1, l = 2√π, Δ = 0.37, option="sBs", indices = 0)

Stabilization of Gottesman-Kitaev-Preskill (GKP) code using the same methods as in Royer et al. (2020) and De Neeve et al. (2022) but to the exception that here the stabilizer measurement outcomes are not discarded but recorded. The simulation is thus a state vector simulation which outputs the bosonic state as a state vector and the measurement outcome. DOI: 10.1038/s41586-023-05782-6

Inputs:

rho::Ket{B} State vector defined in the position/momentum basis B or a composite basis B with at least one position/momentum basis in it.

P::CVsim_Parameters Parameter structure as defined in parameters.jl

n_times::Int Number of stabilization rounds

ε, α, μ ::Float Three parametres as defined DeNeeve_stabilization

option::String String specifying which can contain "q" or "p" for correcting only one of the quadratures. If none, both will be stabilized.

indices Index or vector of indices of subspaces that one would like to stabilize. If index = 0, then all the relevant subspaces will be subjected to n_cycles stabilization.

kraus::Bool Use Kraus operators instead of the unitaries and the spin degree of freedom. Default: true.

Output:

A GKP state as a state vector (Ket)

Outcomes of the stabilizer measurement (Bool)