Wigner quasiprobability of a single CV system state

The Wigner quasiprobability representation of a single continuous-variable (CV) system's state $\rho$ is defined as

\[ W(x,p)=\frac{1}{\pi\hbar}\int_{-\infty}^{+\infty} \, \exp\left(\frac{2i}{\hbar}\,p\,y\right) \, \langle x-y \vert \rho \vert x+y\rangle \, \mathrm{d}y \,.\]

It is evident that this representation is defined by the Fourier transform of the antidiagonals of the density matrix $\rho$ expressed in the position basis. Since CVsim.jl relies on position or momentum representations of CV system states, we can obtain their Wigner quasiprobability through a fast Fourier transform of the columns of their "skewed density matrix". The latter corresponds to the original position density matrix but skewed along one direction,

\[ \begin{pmatrix} \cdot & \cdot & \cdot & \cdot \\ \cdot & \cdot & \cdot & \cdot \\ \cdot & \cdot & \cdot & \cdot \\ \cdot & \cdot & \cdot & \cdot \end{pmatrix} \mapsto \begin{pmatrix} \cdot & \cdot & \cdot & \cdot & 0 & 0 & 0 \\ 0 & \cdot & \cdot & \cdot & \cdot & 0 & 0\\ 0 & 0 & \cdot & \cdot & \cdot & \cdot & 0\\ 0 & 0 & 0 & \cdot & \cdot & \cdot & \cdot \end{pmatrix} \]

Note that the first, third, fifth, etc. columns will contain the desired antidiagonal elements. The even columns will contain in-between antidiagonals and hold mostly redundant information. They could be included but that requires extra computation due to how they straddle across the diagonal. Selecting the odd columns and performing their FFT gives us the desired Wigner quasi-probability as a square matrix of the same dimension as $\rho$.

This calculation has been first implemented in Python by the Wikipedia user Nanite in their "Long-time evolution of a mixed state ρ in an anharmonic potential well." simulation. We adapted it in Julia here.

CVsim.wignerify — Method

wignerify(mat)

Compute the Wigner transform of position-basis density matrix which is a Fourier transform of the anti-diagonals of the matrix (cf. Wikipedia).

Pad the matrix along one dimension, then skew the matrix. This converts the antidiagonals to straight columns.

       |‾‾‾‾|          |‾‾‾‾000|
       |    |    -->   |0    00|
       |    |          |00    0|
       |____|          |000____|

Note that the first, third, fifth, etc. columns will contain the desired antidiagonal elements. The even columns will contain in-between antidiagonals. Here we are interested only in the former ones, i.e. the skew matrix is a square matrix of the same dimensions as the original density matrix in the position basis.

This calculation has been first implemented in Python by the Wikipedia user "Nanite" in their "Long-time evolution of a mixed state ρ in an anharmonic potential well." simulation and adapted in Julia here. src: https://en.wikipedia.org/wiki/File:Hamiltonianflowquantum.webm

Inputs:

mat::Matrix{ComplexF64} : Matrix representing an operator (density matrix) in a position basis.

Output:

Matrix{Float64} : Wigner transform of the matrix (same shape as mat).

In CVsim.jl, we overload this function to enable seamless integration with QuantumOptics.jl, specifically Operator objects (representing density matrices) and StateVector objects (encompassing Ket and Bra states), all defined in a position basis.

CVsim.wignerify — Method

wignerify(rho)
wignerify(rho,indices)

Compute the Wigner quasiprobability distribution for the given operator rho.

If its basis B is a single position basis it returns a single distribution.

If instead B is a composite basis then it will calculate the Wigner quasiprobability distribution for all the subsystems which have a position basis.

Inputs:

rho::Operator{B,B} : Operator (density matrix) defined in the basis B

indices::Vector{Int} : Indices of the subspaces for which you would like a Wigner quasiprobability

Output:

if B == PositionBasis:

Matrix{Float64} : Wigner quasiprobability of dimension n = length(B)

if B == CompositeBasis or if indices are specified:

Vector{Matrix{Float64}} : Vector with Wigner quasiprobability of dimension n_k = length(B_k) where B_k is the position basis of a subspace of B

CVsim.wignerify — Method

wignerify(psi)
wignerify(psi,indices)

Compute the Wigner quasiprobability distribution for the given state vector (i.e. Ket or Bra). The function just transform the state vector to a density matrix and calculate the distribution.

Inputs:

psi::StateVector{B} : State vector (Ket or Bra) defined in the basis B

indices::Vector{Int} : Indices of the subspaces for which you would like a Wigner quasiprobability

Output:

if B == PositionBasis:

Matrix{Float64} : Wigner quasiprobability of dimension n = length(B)

if B == CompositeBasis or if indices are specified:

Vector{Matrix{Float64}} : Vector with k Wigner quasiprobability of dimension n_k = length(B_k) where B_k is the position basis of a subspace of B